fixed test build, fixed bug
bug was that deleting an entity while iterating over
entities would cause an exception to be raised.  This was happening
because all the entities components would be "soft" deleted during iteration
but the entity was actually deleted.  When iteration was over the components
would be deleted for real but they would also try to update the component mask
for the entity they belonged to.

This update wasn't necessary because it was already done during the "soft" delete
but more importantly this caused a crash because that entity no longer had an
entry in the map for it's component mask.

This also includes a regression test for the bug.