Fixed improper values on Component creation
Includes a regression test for this bug. Description from the test:

The situation is that a component is created with certain values and then deleted.
When a new component of the same type is created it was being assigned to the spot
of the old component without having its new values copied into place.